Added a shortcut "-rebuild" flag
This will build the command-line tools and the current binary for the current target, if applicable. It will respect "-debug" and "-clean" to perform only a debug or release rebuild, and to clean before the rebuild if necessary.

In other words, it shortens "pazu rebuild tools,emscripten -release && pazu build emscripten" to a single "pazu build emscripten -rebuild" command

to rebuild the tools and current target and configuration