Use high precision only for vertex shaders.
It's not guaranteed to be available on fragment shaders, but the GLSL ES specification requires it in vertex shaders. Also, vertex shaders are the ones that are likely to need high precision.

I added iOS because Apple (and others) recommend using high precision for vertex shaders. I'd add Blackberry too, but I've never compiled for it.