CanvasBitmap.render now uses __renderTransform in support of scrollRect.
Like the GL renderer, CanvasBitmap needs to subtract out __scrollRect.topLeft before rendering if it intends to use scrollRect functionality without offset.  __renderTransform is already set up for this; it's copied from worldTransform and offset by -__worldOffset, which in turn includes __offset, which comes from __scrollRect.topLeft.

This was part of an attempt to allow the spritesheet library to do subrect rendering, which succeeded and allowed use of the 9-arg version of CanvasRenderContext2D.drawImage, but requires some changes to Bitmap's get_width and get_height and similar to avoid breaking other assumptions, so the rest of this change probably comes somewhat later; in the interim, this should fix basic scrollRect functionality in canvas renderer.

Note that a similar change probably needs to be made in the Cairo backend, and possibly others.

Reviewed in house by Jon Meschino; could definitely use some additional external eyes, especially to make sure it doesn't break any invariants or represent a fundamental misunderstanding of the transform/matrix stack.