console: workaround for nullptr issue with extern classes
This is due to to a change in lime where lime.graphics.ConsoleRenderContext is
implemented with an extern class for some additional type safety, and with less
code. However, it faces an issue with Haxe where `null` does not generate an
actual C/C++ null pointer, and due to the nature of extern classes in Haxe, the
layer of hxcpp which would convert a Haxe null to a C/C++ null is not there.