Allow encode(rect, "png")-style calls
The FlxScreenGrab of HaxeFlixel depends on this.  It doesn't let you pass in a quality, so quality 1 is assumed, but at least it works instead of just returning null, on HaxeFlixel's current default of using OpenFL legacy.  (That's why I left the compatibility comment in there--not sure how it was planned to add that in, without adding an optional parameter, which seems to work well in general except on flash, where it would be a breaking change.  So: partial compatibility without the breaking-change cost.)