EventDispatcherTest: comment out ineffective `useCapture=true` test.
As it stands, this test defaulted to success (`var correctPhase = true;`) and was succeeding, but only because the `listener` was never called.

To go through CAPTURING_PHASE, it's my understanding that events have to go through a `DisplayObject` hierarchy, possibly from a `Stage` instance.  (It's unclear what the behavior is supposed to be re detached `DisplayObject` trees, may be worth testing that and bugging.)

Discovered that this test was ineffective on AS3 and other platforms while looking into how to test #1293 properly.

Submitting this PR as an RFC of sorts; it may want to be merged as a warning of a non-functional test anyway?