Fixed computation of newlines
previously, we checked for ' ' or '-' to detect if a line should be broken.
We should however also check for '\n'.
Otherwise the text 'use\nto' would be seen as 1 word.

In the code we also do an offset of 2px in x and y for the text.
that seems correctly rendered, but the width should be updated accordingly.
( + 2, not - 2 ). This gives better results