Flash fails silently on all errors from ExternalInterface.call(), including function not defined. http://help.adobe.com/en_US/FlashPlatform/reference/actionscript/3/flash/external/ExternalInterface.html#call()
However, now that we're in JS target, there is no concept of
inside/outside of vm. Therefore, it makes sense to forward any
exceptions thrown by the external functions. So we'll allow that,
but we'll continue to ignore any errors about missing functions,
since some [translated] code [likely] relies on that behavior still.