Fix wrong movieclip being created
On android we hit a very weird issue where sometimes the movieclip would be empty (or have the wrong number of children), digging and digging we couldn't figure out why, it didn't makes sense. Turns out that while debugging, I notice that the wrong symbol was used during the movieclip creation but the parent definitely had the correct symbol when it tried to create the child. 

Somehow we must have one of our extension that use thread on android, a callback is called, some movieclip get created at the same time and that static variable get changed.

I think it's much simpler like that and less error prone, was there a reason why it was using static variable?