FileReference's Browse and Load initial support for html5
The code should work.
It is probably not formatted correctly.
I am not sure if the callbacks deserve their own function or should be inlined as they are now.

---

A note on a sister class of this: FileReferenceList;
While we could open the input window and select multiple files, creating multiple FileReferences is hard since we can't just "share" the files that we got with FileReferenceList (or at least I didn't find a way to do so)

---

This is a first approach. Sorry I couldn't clone the repo and made instead a github change proposal.