Fix the `force` logic in `__updateCacheBitmap`
We want to force a cache-bitmap when rendering to canvas with non-default colortransform even if the object is not cacheAsBitmap'd itself, because we have to render to a temporary bitmap and apply color transform to it before rendering it into a canvas, because there's no colorTransform functionality in HTML canvas drawing.

So the force is passed as `true` in these cases, but we didn't check it correctly: the flag should forcefully enable cacheAsBitmap and relevant processing, while it should NOT matter for determining if we need to re-render anything, because that is already handled by __renderDirty (which is also set to `true` when we change colorTransform).