Fix TextField bitmap caching.
Since we changed the meaning of the "force" flag in #107, the value passed in the TextField override became incorrect: it SHOULD NOT force the bitmap caching, but it SHOULD force the re-render if there is bitmap caching. Basically, it's just another dirty flag. And since it's only available in TextField, we gotta use a virtual method to determine the need for a re-render so we can override it in TextField and honor its additional flag.

Also added some TODOs to look into at some point.