Stage: Fixed issue where the html5 target loses focus to the browser every time that Tab or Shift+Tab is pressed
Cancels window.onKeyDown when passing focus to another OpenFL display object, which prevents the browser from switching focus. However, does not cancel window.onKeyDown when wrapping around to the beginning or end.

This change seems to have no effect on other targets, so it doesn't need to be conditionally compiled for html5-only.