Fix for the wrong behavior of HOME/END keys
Incorrect behavior of HOME and END buttons. Currently, whenever HOME/END is pressed, this results in selecting a fragment of text from the current cursor position to the beginning/end of the line. The correct behavior would be just moving the cursor to the beginning/end of the line without selecting any text, whereas text selection should happen when HOME/END is pressed together with Shift. The proposed change fixes this issue.