Unescaping should happen after parsing
HTML unescaping should happen after parsing, not before, because unescaping before parsing defeats the whole purpose of escaping. The purpose of escaping is the prevent the HTML service characters from being parsed as service characters. For example, when I assign a "&amp;gt;" to TextField.htmlText, I'm expecting to see a "&gt;" on the screen, not a ">" (which is the case now).