Prevent null error on MouseEvents when listener closes window
So, for example, if you had a button in a new Window that closed and cleaned up the window on a mouse event listener, the application would crash due to a null object. Ensuring the window is not null before trying to access it on Lines #2539 & #2549 and provides an adequate fix.