TextEngine: strip away the combined width of the whitespace at the end of a line when determining if long words should be broken into multiple lines
The whitespace's width was included in the width of the preceding word, but that makes the word longer than it actually should be. The extra length could cause a break to be added in between the word and the whitespace, which would incorrectly push the whitespace to the beginning of the next line, which we don't want.