TextLayout: Resolve inconsistant letter spacing
Thanks @ShallowMallow for discovering the cause of this! 

I commented out `FT_LOAD_FORCE_AUTOHINT` for now, and the difference in text seems to be greatly improved. I'm not sure if it makes sense at this time to simply get rid of it or if we need to update HB...or some other solution. In any case, I added a TODO to come back to it later. I didn't discover any issues with the change so it seems like a no-brainer to merge considering the favorable impact of text positioning.