TextEngine: Handle scrollV overflow + getter fix
Makes sure scrollV overflow is handled properly. Changed the getter to return `maxScrollV` instead of `maxScrollV - 1`. I'm not sure why I initially changed it to subtract one, but there may have been an inconsistency in the return value of maxScrollV at some point (and there may still be). Best case, it was just an error on my part, but I've reverted it regardless of why the change was made.