Socket: Include missing properties for sys targets
This implements the missing properties for localPort, localAddress, remotePort, and remoteAddress for sys targets.

The js Websocket api does not directly provide the same functionality although there is a way to basically implement it. Considering it is not a crucial component as it is with raw sockets, it may be a little while until we get to implementing the properties for html5.