Graphics: on HiDPI screens, snap to the nearest *device* pixel instead of the nearest *stage* pixel
This code was written before HiDPI screens were common, so it was safe to assume
that device and stage pixels were the same thing. Now, HiDPI screens are much
more common (especially on mobile), so there are usually more device pixels than
stage pixels, and we can improve the snapping behavior.

To review, if pixelRatio is 1, 1 stage pixel == 1 device pixel. If pixelRatio is
2, then 1 stage pixel == 2 device pixels. If pixelRatio is N, then 1 stage pixel
== N device pixels.

Crisp rendering relies on snapping to device pixels, so we can snap to a
fractional stage pixel, as long as multiplying that value by the pixelRatio
gives us an integer device pixel. Not only will we continue to get crisper
rendering from snapping to a whole pixel, but the increased precision will
result in smoother animations as well.