FileReferenceList: dispatch Event.CANCEL if files.length == 0 on html5 (closes #2672)
Same as FileReference. This catches a specific case where no files are selected, but it there are other cases where the file dialog is cancelled, but no event is dispatched, so sometimes, Event.CANCEL is not dispatched by FileReferenceList when it would have been in Flash. I haven't found a workaround for that. We just need to live with it for now. Maybe the web platform will add a cancel event in the future.