EventDispatcher: can't allow useWeakReference on cpp target
The way that hxcpp passes methods by reference to addEventListener() appears to inolve creating another object that wraps the method. There is often only one reference to this wrapper object (which is passed directly to addEventListener), so if a weak reference is used, the wrapper object may be garbage collected, even if the method and its owning object are not garbage collected too.

useWeakReference remains available on html5 target (and flash/air, of course).