ApplicationMain: fix conditional compilation that breaks app initialization when utest is used (closes #2687)
With this conditional compilation, utest cannot be used with the default index.html template, so a custom one was **always** needed. It should just work with the default template.

I'm leaving the munit conditional compilation as-is. I suspect that munit requires this conditional compilation because it doesn't natively support OpenFL (it works with hxml files and doesn't seem to have an option to use project.xml files instead, and it has its own launcher). utest may have been incorrectly included in the conditional compilation at the same time as munit with a wrong assumption that it should work similarly because they're both testing frameworks/libraries.

A user should still be able to create a custom ApplicationMain.hx and index.html to go back to the old behavior. However, I strongly suspect that there's no one who is using utest that needs the old behavior, and that it was just broken.