FileReference: download() html5 improvements
If we allow the download method to be exposed on html5, it should at least do something. This change makes it work minimally even though we can not open a file dialog on all browsers yet. 

Chromes File API is slowly being adopted, so we may be able to support this better in the future.