URLRequest: add withCredentials property for web requests (closes #2719)
https://developer.mozilla.org/en-US/docs/Web/API/XMLHttpRequest/withCredentials

This is a property that didn't exist in Flash, but it is necessary for OpenFL's
HTML5 target to be able to customize the behavior of the internal
XMLHttpRequest in Lime's HTTPRequest class. Uses the HTTPRequest withCredentials
property, which was exposed for similar reasons.

At some point in the past, URLRequest's manageCookies property was incorrectly
used to set HTTPRequest's withCredentials property. However, manageCookies and
withCredentials are not the same thing. They both control the behavior of
cookies, but under different circumstances. manageCookies is supposed to control
whether cookies are ever saved and passed with requests. withCredentials
specifically applies to cross-origin requests only.

Setting withCredentials will have no affect on targets that don't run in web
browsers because all requests on native are essentially treated as same-origin.