Graphics: fix using hardware-only BitmapData with beginBitmapFill() and drawTriangles()
In commit c54052d5fa9049ab4bee54bfa39823fe72c773af, beginBitmapFill() was made to divert to beginFill() with black if a BitmapData was hardware-only. This is because many drawing APIs (like drawRect()) are software (Canvas or Cairo) only, and nothing would render. It would also lead to hit testing issues. However, because drawTriangles() (and possibly others) can be used with hardware-only BitmapData, this broke some workflows. Instead, beginBitmapFill() will go back to the previous behavior that allowed hardware drawing, but if it needs to go to software instead, it will draw as a black rectangle like in the previously mentioned commit. Basically, this is a better fix for the original issue (references #2646), while restoring the ability to draw hardware-only BitmapData in hardware.