BitmapData: Dispose BaseTexture if not shared
Ideally we dispose a texture if its not being shared.

We can reasonably predict when a BitmapData's Textured is shared or not and dispose of it. Typically this means we called getTexture(), clone() or fromTexture().

With this change, we track when the Texture is being shared and then proceed to call dispose on the texture when the BitmapData is disposed.

I'm not 100% fond of this solution but we need to prevent the memory leak caused by the texture failing to be disposed.

Open to other ideas or solutions.