BitmapData: Documentation
So, lock and unlock do absolutely nothing on every target except for Flash/AIR at the moment. I really don't like having class methods that don't do anything, but it's even worse not to inform users of that in the documentation. 

Even I assumed this provided some sort of optimization here, atleast in software rendering... Haha.