actions: add -Duse-lime-tools to fix the Haxe 3 and Neko unit tests
Something is now using ThreadPool, which requires Application.current to be non-null, but the hxp version of the tests doesn't set up an Application. So now, all unit tests on GitHub Actions will use Lime tools, which sets up the Application automatically. It may be possible to rewrite the tests to allow them in hxp, but I'm not experienced enough with hxp to do it, so someone else can take that on, if they find it valuable.